refactor: migrate inventory feature to feature slice
- Move InventoryContext and InventoryProvider to features/inventory/
- Move useInventory hook to features/inventory/hooks/
- Move status calculation utilities (getItemStatus, isItemExpired,
  getDaysUntilExpiration, etc.) to features/inventory/utils/
- Move inventory components (ItemCard, ItemList, ItemForm,
  CategoryNav, FilterBar, CategoryStatusSummary) to
  features/inventory/components/
- Create public API index.ts for clean imports
- Create backward-compatible re-exports in shared/ and
  components/inventory/ for gradual migration
- Update tests and fix itemType requirement in ItemList test

The inventory feature follows the same pattern as alerts, templates,
categories, and household features. All components, hooks, utils,
and context are now organized under features/inventory/.

Import from '@/features/inventory' instead of individual files.
