+++ src/components/inventory/FilterBar.tsx
@@ -1,72 +1,6 @@
-import { useTranslation } from 'react-i18next';
-import { Select } from '@/shared/components/Select';
-import { Input } from '@/shared/components/Input';
-import type { ItemStatus } from '@/shared/types';
-import styles from './FilterBar.module.css';
-
-export interface FilterBarProps {
-  searchQuery: string;
-  onSearchChange: (query: string) => void;
-  statusFilter: ItemStatus | 'all';
-  onStatusFilterChange: (status: ItemStatus | 'all') => void;
-  sortBy: 'name' | 'quantity' | 'expiration';
-  onSortByChange: (sortBy: 'name' | 'quantity' | 'expiration') => void;
-}
-
-export const FilterBar = ({
-  searchQuery,
-  onSearchChange,
-  statusFilter,
-  onStatusFilterChange,
-  sortBy,
-  onSortByChange,
-}: FilterBarProps) => {
-  const { t } = useTranslation();
-
-  const statusOptions = [
-    { value: 'all', label: t('inventory.filter.allStatuses') },
-    { value: 'ok', label: t('status.ok') },
-    { value: 'warning', label: t('status.warning') },
-    { value: 'critical', label: t('status.critical') },
-  ];
-
-  const sortOptions = [
-    { value: 'name', label: t('inventory.sort.name') },
-    { value: 'quantity', label: t('inventory.sort.quantity') },
-    { value: 'expiration', label: t('inventory.sort.expiration') },
-  ];
-
-  return (
-    <div className={styles.filterBar}>
-      <div className={styles.search}>
-        <Input
-          type="text"
-          value={searchQuery}
-          onChange={(e) => onSearchChange(e.target.value)}
-          placeholder={t('inventory.searchPlaceholder')}
-          aria-label={t('inventory.searchLabel')}
-        />
-      </div>
-
-      <div className={styles.filters}>
-        <Select
-          value={statusFilter}
-          onChange={(e) =>
-            onStatusFilterChange(e.target.value as ItemStatus | 'all')
-          }
-          options={statusOptions}
-          label={t('inventory.filter.status')}
-        />
-
-        <Select
-          value={sortBy}
-          onChange={(e) =>
-            onSortByChange(e.target.value as 'name' | 'quantity' | 'expiration')
-          }
-          options={sortOptions}
-          label={t('inventory.sort.label')}
-        />
-      </div>
-    </div>
-  );
-};
+/**
+ * @deprecated Use @/features/inventory instead
+ * This file is kept for backward compatibility during migration.
+ */
+export { FilterBar } from '@/features/inventory';
+export type { FilterBarProps } from '@/features/inventory';
