+++ src/features/inventory/components/CategoryNav.module.css
