+++ src/features/inventory/components/CategoryNav.stories.tsx
