+++ src/features/inventory/components/CategoryNav.test.tsx
