+++ src/features/inventory/components/CategoryStatusSummary.module.css
