+++ src/features/inventory/components/CategoryStatusSummary.test.tsx
