+++ src/features/inventory/components/FilterBar.stories.tsx
