+++ src/features/inventory/components/ItemCard.module.css
