+++ src/features/inventory/components/ItemCard.test.tsx
