+++ src/features/inventory/components/index.ts
@@ -1,0 +1,20 @@
+export { ItemCard } from './ItemCard';
+export type { ItemCardProps } from './ItemCard';
+
+export { ItemList } from './ItemList';
+export type { ItemListProps } from './ItemList';
+
+export { ItemForm } from './ItemForm';
+export type { ItemFormProps } from './ItemForm';
+
+export { CategoryNav } from './CategoryNav';
+export type { CategoryNavProps } from './CategoryNav';
+
+export { FilterBar } from './FilterBar';
+export type { FilterBarProps } from './FilterBar';
+
+export { CategoryStatusSummary } from './CategoryStatusSummary';
+export type {
+  CategoryStatusSummaryProps,
+  CategoryShortage,
+} from './CategoryStatusSummary';
