+++ src/features/inventory/hooks/useInventory.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { InventoryContext } from '../context';
+
+export function useInventory() {
+  const context = useContext(InventoryContext);
+  if (!context) {
+    throw new Error('useInventory must be used within InventoryProvider');
+  }
+  return context;
+}
