+++ src/features/inventory/index.ts
@@ -1,0 +1,40 @@
+// Context
+export { InventoryContext } from './context';
+export type { InventoryContextValue } from './context';
+
+// Provider
+export { InventoryProvider } from './provider';
+
+// Hooks
+export { useInventory } from './hooks';
+
+// Utils
+export {
+  getDaysUntilExpiration,
+  isItemExpired,
+  getItemStatus,
+  calculateItemStatus,
+  getStatusFromPercentage,
+  getStatusFromScore,
+  getStatusVariant,
+} from './utils';
+
+// Components
+export {
+  ItemCard,
+  ItemList,
+  ItemForm,
+  CategoryNav,
+  FilterBar,
+  CategoryStatusSummary,
+} from './components';
+
+export type {
+  ItemCardProps,
+  ItemListProps,
+  ItemFormProps,
+  CategoryNavProps,
+  FilterBarProps,
+  CategoryStatusSummaryProps,
+  CategoryShortage,
+} from './components';
