+++ src/features/inventory/types.ts
@@ -1,0 +1,9 @@
+/**
+ * Inventory feature types
+ *
+ * Core types (InventoryItem, Category, etc.) are in @/shared/types
+ * This file re-exports the context value type for convenience.
+ */
+
+// Re-export context value type for convenience
+export type { InventoryContextValue } from './context';
