+++ src/features/inventory/utils/index.ts
@@ -1,7 +1,3 @@
-/**
- * @deprecated Use @/features/inventory instead
- * This file is kept for backward compatibility during migration.
- */
 export {
   getDaysUntilExpiration,
   isItemExpired,
@@ -10,4 +6,4 @@ export {
   getStatusFromPercentage,
   getStatusFromScore,
   getStatusVariant,
-} from '@/features/inventory';
+} from './status';
