+++ src/features/inventory/utils/status.test.ts
