+++ src/shared/hooks/useInventory.ts
@@ -1,10 +1,5 @@
-import { useContext } from 'react';
-import { InventoryContext } from '@/shared/contexts/InventoryContext';
-
-export function useInventory() {
-  const context = useContext(InventoryContext);
-  if (!context) {
-    throw new Error('useInventory must be used within InventoryProvider');
-  }
-  return context;
-}
+/**
+ * @deprecated Use @/features/inventory instead
+ * This file is kept for backward compatibility during migration.
+ */
+export { useInventory } from '@/features/inventory';
